+++ android/app/src/main/AndroidManifest.xml
@@ -80,7 +80,9 @@
     <application
         android:label="DOCTO"
         android:name="${applicationName}"
-        android:icon="@mipmap/ic_launcher">
+        android:icon="@mipmap/ic_launcher"
+        android:allowBackup="false"
+        android:fullBackupContent="false">
         <meta-data android:name="com.google.android.geo.API_KEY"
                android:value="${mapsApiKey}"/>
         <activity
