+++ lib/constants.dart
@@ -1,2 +1,14 @@
 const String healthSdkSharedPreferenceKey='healthSdkValue';
 const String isOnBoardingSharedPreferenceKey='isOnBoardingComplete';
+
+class PrefKeys {
+  static const String userName = 'userNameSharedPreferenceKey';
+  static const String profileImagePath = 'userProfileImagePathSharedPreferenceKey';
+  static const String weight = 'userWeightSharedPreferenceKey';
+  static const String weightUnit = 'userWeightUnitSharedPreferenceKey';
+  static const String height = 'userHeightSharedPreferenceKey';
+  static const String heightUnit = 'userHeightUnitSharedPreferenceKey';
+  static const String gender = 'userGenderSharedPreferenceKey';
+  static const String healthGoal = 'userHealthGoalSharedPreferenceKey';
+  static const String age = 'userAgeSharedPreferenceKey';
+}
